Rename old implementation
